CMS added collection support

+++ src/My/CoreBundle/Entity/File.php
@@ -4,6 +4,7 @@ namespace My\CoreBundle\Entity;
 
 use Doctrine\ORM\Mapping as ORM;
 use Symfony\Component\Validator\Constraints as Assert;
+use Gedmo\Mapping\Annotation as Gedmo;
 use Symfony\Component\HttpFoundation\File\UploadedFile;
 
 /**
@@ -23,11 +24,34 @@ abstract class File
     private $id;
 
     /**
+     * @Gedmo\Timestampable(on="create")
+     * @ORM\Column(type="datetime")
+     */
+    private $createdAt;
+
+    /**
+     * @ORM\Column(type="string", length=255)
+     * @Assert\NotNull()
+     * @Assert\Length(max="255")
+     */
+    private $originalName;
+
+    /**
      * @ORM\Column(type="string", length=255)
      * @Assert\NotNull()
      * @Assert\Length(max="255")
      */
     private $path;
+
+    /**
+     * @ORM\Column(type="integer")
+     */
+    private $size;
+
+    /**
+     * @ORM\Column(type="string", length=255)
+     */
+    private $mimeType;
 
     /**
      * @Assert\File(maxSize="6000000")
@@ -59,6 +83,8 @@ abstract class File
     }
 
     /**
+     * Pre upload
+     *
      * @ORM\PrePersist()
      * @ORM\PreUpdate()
      */
@@ -66,7 +92,11 @@ abstract class File
     {
         if (null !== $this->getFile()) {
             $filename = sha1(uniqid(mt_rand(), true));
-            $this->path = $filename.'.'.$this->getFile()->guessExtension();
+
+            $this->setOriginalName($this->getFile()->getClientOriginalName());
+            $this->setPath($filename.'.'.$this->getFile()->getClientOriginalExtension());
+            $this->setSize($this->getFile()->getClientSize());
+            $this->setMimeType($this->getFile()->getMimeType());
         }
     }
 
@@ -82,7 +112,7 @@ abstract class File
 
         $this->getFile()->move($this->getUploadRootDir(), $this->path);
 
-        if (isset($this->temp)) {
+        if (isset($this->temp) && file_exists($this->getUploadRootDir().'/'.$this->temp)) {
             unlink($this->getUploadRootDir().'/'.$this->temp);
             $this->temp = null;
         }
@@ -156,5 +186,97 @@ abstract class File
     public function getPath()
     {
         return $this->path;
+    }
+
+    /**
+     * Set createdAt
+     *
+     * @param \DateTime $createdAt
+     * @return File
+     */
+    public function setCreatedAt($createdAt)
+    {
+        $this->createdAt = $createdAt;
+
+        return $this;
+    }
+
+    /**
+     * Get createdAt
+     *
+     * @return \DateTime
+     */
+    public function getCreatedAt()
+    {
+        return $this->createdAt;
+    }
+
+    /**
+     * Set size
+     *
+     * @param integer $size
+     * @return File
+     */
+    public function setSize($size)
+    {
+        $this->size = $size;
+
+        return $this;
+    }
+
+    /**
+     * Get size
+     *
+     * @return integer
+     */
+    public function getSize()
+    {
+        return $this->size;
+    }
+
+    /**
+     * Set mimeType
+     *
+     * @param string $mimeType
+     * @return File
+     */
+    public function setMimeType($mimeType)
+    {
+        $this->mimeType = $mimeType;
+
+        return $this;
+    }
+
+    /**
+     * Get mimeType
+     *
+     * @return string
+     */
+    public function getMimeType()
+    {
+        return $this->mimeType;
+    }
+
+    /**
+     * Set originalName
+     *
+     * @param string $originalName
+     * @return File
+     */
+    public function setOriginalName($originalName)
+    {
+        $this->originalName = $originalName;
+
+        return $this;
+    }
+
+    /**
+     * Get originalName
+     *
+     * @return string
+     */
+    public function getOriginalName()
+    {
+        return $this->originalName;
     }
 }
